feat: implement SignatureFieldRenderer component for enhanced signature handling
