+++ components/docs/forms/FieldRenderer.tsx
@@ -9,7 +9,6 @@
 
 "use client";
 
-import { useSignContext } from "@/app/docs/auth/provider/sign.ctx";
 import {
   FormCheckbox,
   FormDatePicker,
@@ -20,19 +19,9 @@ import {
 } from "./EditForm";
 import { AutocompleteTreeMulti, TreeOption } from "./autocomplete";
 import { ClientField } from "@betterinternship/core/forms";
-import { useEffect, useRef, useState } from "react";
-import { Eye, ImageUp, PenLine, Trash2, Type, UploadCloud } from "lucide-react";
-import { removeSignatureImageBackground } from "@/lib/signature-image-cleanup";
-import {
-  createSignatureImageValue,
-  getSignatureImageFieldKey,
-  parseSignatureImageValue,
-  serializeSignatureImageValue,
-  type SignatureImageValue,
-} from "@betterinternship/core/forms";
-
-const MAX_SIGNATURE_UPLOAD_BYTES = 10 * 1024 * 1024;
-const MAX_SIGNATURE_UPLOAD_LABEL = "10 MB";
+import { useState } from "react";
+import { Eye } from "lucide-react";
+import { SignatureFieldRenderer } from "./SignatureFieldRenderer";
 
 export const FieldRenderer = <T extends any[]>({
   field,
@@ -150,7 +139,7 @@ export const FieldRenderer = <T extends any[]>({
   // Signatures
   if (field.type === "signature") {
     return (
-      <FieldRendererSignature
+      <SignatureFieldRenderer
         field={field}
         value={value}
         TooltipContent={TooltipLabel}
@@ -423,395 +412,6 @@ const FieldRendererInput = <T extends any[]>({
         onBlur={() => onBlur?.()}
         labelAddon={badge}
       />
-      <TooltipContent />
-    </div>
-  );
-};
-
-/**
- * Signature-specific input
- *
- * @component
- */
-const FieldRendererSignature = <T extends any[]>({
-  field,
-  value,
-  TooltipContent,
-  onChange,
-  onAuxValueChange,
-  onBlur,
-  allValues = {},
-}: {
-  field: ClientField<T>;
-  value: string;
-  TooltipContent: () => React.ReactNode;
-  onChange: (v: string | number) => void;
-  onAuxValueChange?: (key: string, value: any) => void;
-  onBlur?: (nextValue?: unknown) => void;
-  allValues?: Record<string, string>;
-}) => {
-  const signContext = useSignContext();
-  const [checked, setChecked] = useState(false);
-  const imageFieldKey = getSignatureImageFieldKey(field.field);
-  const currentSignatureImage = parseSignatureImageValue(allValues[imageFieldKey]);
-  const [mode, setMode] = useState<"type" | "upload" | "draw">(
-    currentSignatureImage?.source === "draw"
-      ? "draw"
-      : currentSignatureImage?.source === "upload"
-        ? "upload"
-        : "type"
-  );
-  const [typedName, setTypedName] = useState(value || "");
-  const [uploadError, setUploadError] = useState("");
-  const [isUploadDragging, setIsUploadDragging] = useState(false);
-  const canvasRef = useRef<HTMLCanvasElement | null>(null);
-  const isDrawingRef = useRef(false);
-
-  // ! PUT THIS SOMEWHERE ELSE
-  useEffect(() => {
-    signContext.setHasAgreedForSignature(field.field, value, checked);
-  }, [checked, value]);
-
-  useEffect(() => {
-    const parsed = parseSignatureImageValue(allValues[imageFieldKey]);
-    if (parsed) {
-      setMode(parsed.source);
-    }
-
-    setTypedName(value || "");
-  }, [allValues, imageFieldKey, value]);
-
-  const emitSignatureImage = (nextImage: SignatureImageValue) => {
-    onAuxValueChange?.(imageFieldKey, serializeSignatureImageValue(nextImage));
-  };
-
-  const getSignatureImageSrc = (signature: SignatureImageValue | null) => {
-    if (!signature) return "";
-    if (signature.image.storage === "bucket") {
-      return signature.image.signedUrl || signature.image.publicUrl || "";
-    }
-    return signature.image.dataUrl;
-  };
-
-  const clearCanvas = () => {
-    const canvas = canvasRef.current;
-    const ctx = canvas?.getContext("2d");
-    if (!canvas || !ctx) return;
-
-    ctx.clearRect(0, 0, canvas.width, canvas.height);
-  };
-
-  const drawImageOnCanvas = (dataUrl: string) => {
-    const canvas = canvasRef.current;
-    const ctx = canvas?.getContext("2d");
-    if (!canvas || !ctx) return;
-
-    const image = new Image();
-    image.onload = () => {
-      clearCanvas();
-      const scale = Math.min(canvas.width / image.width, canvas.height / image.height);
-      const width = image.width * scale;
-      const height = image.height * scale;
-      const x = (canvas.width - width) / 2;
-      const y = (canvas.height - height) / 2;
-      ctx.drawImage(image, x, y, width, height);
-    };
-    image.src = dataUrl;
-  };
-
-  useEffect(() => {
-    if (mode !== "draw") return;
-    window.requestAnimationFrame(() => {
-      const latestImage = parseSignatureImageValue(allValues[imageFieldKey]);
-      if (latestImage?.source === "draw") {
-        const src = getSignatureImageSrc(latestImage);
-        if (!src) {
-          clearCanvas();
-          return;
-        }
-        drawImageOnCanvas(src);
-        return;
-      }
-      clearCanvas();
-    });
-  }, [allValues, imageFieldKey, mode]);
-
-  const handleTypedNameChange = (nextName: string) => {
-    setTypedName(nextName);
-    onChange(nextName);
-  };
-
-  const clearSignatureImage = () => {
-    setUploadError("");
-    clearCanvas();
-    onAuxValueChange?.(imageFieldKey, "");
-  };
-
-  const handleUpload = (file: File | undefined) => {
-    if (!file) return;
-
-    if (file.type !== "image/png" && file.type !== "image/jpeg") {
-      setUploadError("Please upload a PNG or JPEG signature image.");
-      return;
-    }
-    if (file.size > MAX_SIGNATURE_UPLOAD_BYTES) {
-      setUploadError(`Signature image must be ${MAX_SIGNATURE_UPLOAD_LABEL} or smaller.`);
-      return;
-    }
-
-    setUploadError("");
-    const reader = new FileReader();
-    reader.onload = async () => {
-      const dataUrl = typeof reader.result === "string" ? reader.result : "";
-      if (!dataUrl) return;
-      const normalizedDataUrl = await removeSignatureImageBackground(dataUrl);
-      emitSignatureImage(
-        createSignatureImageValue({
-          source: "upload",
-          dataUrl: normalizedDataUrl,
-          mimeType: "image/png",
-        })
-      );
-      setMode("upload");
-    };
-    reader.onerror = () => {
-      setUploadError("Unable to read signature image.");
-    };
-    reader.readAsDataURL(file);
-  };
-
-  const exportCanvasSignature = () => {
-    const canvas = canvasRef.current;
-    if (!canvas) return;
-
-    emitSignatureImage(
-      createSignatureImageValue({
-        source: "draw",
-        dataUrl: canvas.toDataURL("image/png"),
-        mimeType: "image/png",
-      })
-    );
-  };
-
-  const getCanvasPoint = (event: React.PointerEvent<HTMLCanvasElement>) => {
-    const canvas = canvasRef.current;
-    if (!canvas) return null;
-
-    const rect = canvas.getBoundingClientRect();
-    return {
-      x: ((event.clientX - rect.left) / rect.width) * canvas.width,
-      y: ((event.clientY - rect.top) / rect.height) * canvas.height,
-    };
-  };
-
-  const handleDrawStart = (event: React.PointerEvent<HTMLCanvasElement>) => {
-    const canvas = canvasRef.current;
-    const ctx = canvas?.getContext("2d");
-    const point = getCanvasPoint(event);
-    if (!canvas || !ctx || !point) return;
-
-    event.currentTarget.setPointerCapture(event.pointerId);
-    isDrawingRef.current = true;
-    ctx.strokeStyle = "#111827";
-    ctx.lineWidth = 4;
-    ctx.lineCap = "round";
-    ctx.lineJoin = "round";
-    ctx.beginPath();
-    ctx.moveTo(point.x, point.y);
-  };
-
-  const handleDrawMove = (event: React.PointerEvent<HTMLCanvasElement>) => {
-    if (!isDrawingRef.current) return;
-    const ctx = canvasRef.current?.getContext("2d");
-    const point = getCanvasPoint(event);
-    if (!ctx || !point) return;
-
-    ctx.lineTo(point.x, point.y);
-    ctx.stroke();
-  };
-
-  const handleDrawEnd = (event: React.PointerEvent<HTMLCanvasElement>) => {
-    if (!isDrawingRef.current) return;
-
-    isDrawingRef.current = false;
-    event.currentTarget.releasePointerCapture(event.pointerId);
-    exportCanvasSignature();
-  };
-
-  const signatureImage = parseSignatureImageValue(allValues[imageFieldKey]);
-  const ModeButton = ({
-    active,
-    children,
-    onClick,
-  }: {
-    active: boolean;
-    children: React.ReactNode;
-    onClick: () => void;
-  }) => (
-    <button
-      type="button"
-      className={`inline-flex h-8 items-center gap-1.5 rounded-[0.33em] border px-2.5 text-xs font-medium transition-colors ${
-        active
-          ? "border-slate-900 bg-slate-900 text-white"
-          : "border-slate-200 bg-white text-slate-700 hover:bg-slate-50"
-      }`}
-      onClick={onClick}
-    >
-      {children}
-    </button>
-  );
-
-  return (
-    <div className="space-y-1.5 rounded-[0.33em] border border-gray-300 p-4 px-5">
-      <FormInput
-        required={true}
-        label={`${field.label} (Signatory Full Name)`}
-        value={typedName ?? ""}
-        setter={handleTypedNameChange}
-        tooltip={field.tooltip_label}
-        className="w-full"
-        onBlur={() => onBlur?.(value)}
-      />
-      <div className="mt-3 flex flex-wrap items-center gap-2">
-        <ModeButton
-          active={mode === "type"}
-          onClick={() => {
-            setMode("type");
-            clearSignatureImage();
-          }}
-        >
-          <Type className="h-3.5 w-3.5" />
-          Type
-        </ModeButton>
-        <ModeButton
-          active={mode === "upload"}
-          onClick={() => {
-            setMode("upload");
-            setUploadError("");
-          }}
-        >
-          <ImageUp className="h-3.5 w-3.5" />
-          Upload
-        </ModeButton>
-        <ModeButton
-          active={mode === "draw"}
-          onClick={() => {
-            setMode("draw");
-            setUploadError("");
-          }}
-        >
-          <PenLine className="h-3.5 w-3.5" />
-          Draw
-        </ModeButton>
-        {signatureImage ? (
-          <button
-            type="button"
-            className="inline-flex h-8 w-8 items-center justify-center rounded-[0.33em] border border-red-200 bg-white text-red-600 transition-colors hover:bg-red-50"
-            onClick={clearSignatureImage}
-            title="Clear signature image"
-            aria-label="Clear signature image"
-          >
-            <Trash2 className="h-3.5 w-3.5" />
-          </button>
-        ) : null}
-      </div>
-      {uploadError ? <p className="mt-2 text-xs text-red-600">{uploadError}</p> : null}
-      {mode === "upload" ? (
-        <label
-          className={`mt-3 flex min-h-36 cursor-pointer flex-col items-center justify-center gap-3 rounded-[0.33em] border border-dashed p-4 text-center transition-colors ${
-            uploadError
-              ? "border-red-300 bg-red-50/60"
-              : isUploadDragging
-                ? "border-slate-900 bg-slate-100"
-                : "border-slate-300 bg-white hover:border-slate-400 hover:bg-slate-50"
-          }`}
-          onDragEnter={(event) => {
-            event.preventDefault();
-            event.stopPropagation();
-            setIsUploadDragging(true);
-          }}
-          onDragOver={(event) => {
-            event.preventDefault();
-            event.stopPropagation();
-            setIsUploadDragging(true);
-          }}
-          onDragLeave={(event) => {
-            event.preventDefault();
-            event.stopPropagation();
-            if (!event.currentTarget.contains(event.relatedTarget as Node | null)) {
-              setIsUploadDragging(false);
-            }
-          }}
-          onDrop={(event) => {
-            event.preventDefault();
-            event.stopPropagation();
-            setIsUploadDragging(false);
-            handleUpload(event.dataTransfer.files?.[0]);
-          }}
-        >
-          {signatureImage && getSignatureImageSrc(signatureImage) ? (
-            <img
-              src={getSignatureImageSrc(signatureImage)}
-              alt="Uploaded signature"
-              className="max-h-32 w-full object-contain"
-            />
-          ) : (
-            <>
-              <span className="flex h-10 w-10 items-center justify-center rounded-full border border-slate-200 bg-slate-50 text-slate-700">
-                <UploadCloud className="h-5 w-5" />
-              </span>
-              <span className="text-sm font-medium text-slate-900">Choose a signature image</span>
-              <span className="text-xs text-slate-500">
-                PNG or JPEG, {MAX_SIGNATURE_UPLOAD_LABEL} maximum
-              </span>
-            </>
-          )}
-          <input
-            type="file"
-            accept="image/png,image/jpeg"
-            className="hidden"
-            onChange={(event) => {
-              handleUpload(event.target.files?.[0]);
-              event.target.value = "";
-            }}
-          />
-        </label>
-      ) : null}
-      {mode === "draw" ? (
-        <div className="mt-3 space-y-2">
-          <canvas
-            ref={canvasRef}
-            width={720}
-            height={220}
-            className="h-36 w-full touch-none rounded-[0.33em] border border-slate-300 bg-white"
-            onPointerDown={handleDrawStart}
-            onPointerMove={handleDrawMove}
-            onPointerUp={handleDrawEnd}
-            onPointerCancel={() => {
-              isDrawingRef.current = false;
-            }}
-          />
-          <button
-            type="button"
-            className="rounded-[0.33em] border border-slate-200 bg-white px-2.5 py-1.5 text-xs font-medium text-slate-700 transition-colors hover:bg-slate-50"
-            onClick={() => {
-              clearCanvas();
-              clearSignatureImage();
-            }}
-          >
-            Clear drawing
-          </button>
-        </div>
-      ) : null}
-      <div className="mt-5 flex flex-row" onClick={() => setChecked(!checked)}>
-        <div className="mt-1 mr-2">
-          <FormCheckbox checked={checked} setter={setChecked}></FormCheckbox>
-        </div>
-        <span className="text-md text-gray-700 italic">
-          I agree to use electronic representation of my signature for all purposes when I (or my
-          agent) use them on documents, including legally binding contracts.
-        </span>
-      </div>
       <TooltipContent />
     </div>
   );
